chore: update OS X defaults
com.apple.Terminal: Set Homebrew as the default and startup profile
com.apple.Dock: Set the icon size small and hot-corners for sleeping the
display and showing the desktop.
com.apple.finder: Keep folders on top when sorting.
com.apple.menuextra.clock: Display the Day and Month on the clock